chore: Updated Z85 encoder.

+++ z85/z85.go
@@ -20,16 +20,19 @@
 //
 // Author: Frank Schwab
 //
-// Version: 3.1.0
+// Version: 3.2.0
 //
 // Change history:
 //    2025-01-31: V1.0.0: Created.
 //    2025-02-01: V2.0.0: Return an error on invalid padding for DecodeAndUnpad.
 //    2025-02-02: V3.0.0: Structured errors.
 //    2025-02-05: V3.1.0: Streamlined result size calculation.
+//    2025-02-05: V3.2.0: Use unsigned types where possible.
 //
 
 // Package z85 implements Z85 encoding as specified in https://rfc.zeromq.org/spec/32.
+// There are two additional functions that implement a padded variant for sources
+// of lengths that are not a multiple of 4. Theses are _not_ standardized.
 package z85
 
 import (
@@ -45,7 +48,7 @@ const codeSize = 85
 // byteChunkSize is the size of a byte chunk to be processed.
 const byteChunkSize = 4
 
-// byteChunkMask Mask to check a byte chunk index.
+// byteChunkMask is the mask to check a byte chunk index.
 const byteChunkMask = byteChunkSize - 1
 
 // byteChunkShift is the shift value used for division by shifting.
@@ -62,7 +65,7 @@ var encodeTable = `0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXY
 // Encode encodes a byte slice into a Z85 encoded string.
 // The length of the slice must be a multiple of 4.
 func Encode(source []byte) (string, error) {
-	sourceLen := len(source)
+	sourceLen := uint(len(source))
 
 	if sourceLen > math.MaxInt/encodedChunkSize {
 		return ``, ErrTooLong
@@ -75,7 +78,7 @@ func Encode(source []byte) (string, error) {
 	chunkCount := sourceLen >> byteChunkShift
 	result := make([]byte, sourceLen+chunkCount)
 	destination := result
-	for chunkIndex := 0; chunkIndex < chunkCount; chunkIndex++ {
+	for chunkIndex := uint(0); chunkIndex < chunkCount; chunkIndex++ {
 		value := binary.BigEndian.Uint32(source[:byteChunkSize])
 
 		// Generate 5 characters
